DW-407: Changed stuff after code review
